Move custom modules to a new lib folder
