+++ lib/connection.js
