+++ lib/query.js
@@ -1,6 +1,6 @@
 var connection = require("./connection");
-var seniorities = require('./models/seniorities');
-var hackers = require('./models/hackers');
+var seniorities = require('../models/seniorities');
+var hackers = require('../models/hackers');
 
 exports.getSeniority = getOne(seniorities);
 exports.getSeniorities = getAll(seniorities);
