+++ lib/response.js
@@ -21,5 +21,5 @@ exports.make = function(req, res, data, template) {
  * @param {string} file
  */
 function getTemplate (fileName, callback) {
-  fs.readFile(__dirname + '/templates/' + fileName, { encoding: 'utf8' }, callback);
+  fs.readFile('./templates/' + fileName, { encoding: 'utf8' }, callback);
 }
