+++ lib/template.js
