+++ lib/utils.js
@@ -1,8 +1,7 @@
 const SKILL_LEVELS=['Nice to have', 'Required', 'Experience', 'Advanced']
 
 // hardcoded stuff
-var hackers = require('./data/json/hackers.json');
-var skills = require('./data/json/skills.json');
+var skills = require('../data/json/skills.json');
 
 module.exports = {
   hackerNeededSkills: hackerNeededSkills,
